firefly/requirements.txt
Mezzanine
django-location-field
django-admin-tools
django-hitcount
django-durationfield
mutagen
django-widget-tweaks